requirements.txt
django
djangorestframework
requests
bs4
lxml
